+++ assets/javascripts/kitten/components/graphics/illustrations/illustrations.stories.js
@@ -9,6 +9,15 @@ import { MixIllustration } from './mix-illustration'
 import { PhoneIllustration } from './phone-illustration'
 import { RocketIllustration } from './rocket-illustration'
 import { KkbbFlashIllustration } from './kkbb-flash-illustration'
+import { HourglassIllustration } from './hourglass-illustration'
+import { HandIllustration } from './hand-illustration'
+import { FormIllustration } from './form-illustration'
+import { ComputerIllustration } from './computer-illustration'
+import { RibbonIllustration } from './ribbon-illustration'
+import { HandAndFlowerIllustration } from './hand-and-flower-illustration'
+import { RainbowIllustration } from './rainbow-illustration'
+import { PencilIllustration } from './pencil-illustration'
+import { PatchNewIllustration } from './patch-new-illustration'
 
 const Wrapper = styled.div`
   padding: ${pxToRem(40)};
@@ -25,8 +34,8 @@ const Group = styled(({ className, children, title }) => {
   )
 })`
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(${pxToRem(140)}, 1fr));
-  grid-gap: ${pxToRem(80)} ${pxToRem(40)};
+  grid-template-columns: repeat(auto-fill, minmax(${pxToRem(350)}, 1fr));
+  grid-gap: ${pxToRem(80)} ${pxToRem(20)};
   margin-bottom: ${pxToRem(50)};
 `
 
@@ -53,8 +62,8 @@ const IllustrationWrapper = styled(({ className, children, name }) => {
   svg {
     width: auto;
     height: auto;
-    max-width: ${pxToRem(150)};
-    max-height: ${pxToRem(160)};
+    max-width: ${pxToRem(250)};
+    max-height: ${pxToRem(300)};
   }
 `
 
@@ -88,6 +97,40 @@ export const AllIllustrations = () => {
       <Group>
         <IllustrationWrapper name="KkbbFlashIllustration">
           <KkbbFlashIllustration size="small" />
+        </IllustrationWrapper>
+      </Group>
+      <Group>
+      <IllustrationWrapper name="HourglassIllustration">
+        <HourglassIllustration />
+      </IllustrationWrapper>
+      <IllustrationWrapper name="HandIllustration">
+        <HandIllustration />
+      </IllustrationWrapper>
+      
+      <IllustrationWrapper name="ComputerIllustration">
+        <ComputerIllustration />
+      </IllustrationWrapper>
+      <IllustrationWrapper name="FormIllustration">
+        <FormIllustration />
+      </IllustrationWrapper>
+      </Group>
+      <Group>
+      <IllustrationWrapper name="PencilIllustration">
+        <PencilIllustration />
+      </IllustrationWrapper>
+      <IllustrationWrapper name="RainbowIllustration">
+        <RainbowIllustration />
+      </IllustrationWrapper>
+      <IllustrationWrapper name="HandAndFlowerIllustration">
+        <HandAndFlowerIllustration />
+      </IllustrationWrapper>
+      <IllustrationWrapper name="RibbonIllustration">
+        <RibbonIllustration />
+      </IllustrationWrapper>
+      </Group>
+      <Group>
+        <IllustrationWrapper name="PatchNewIllustration">
+          <PatchNewIllustration />
         </IllustrationWrapper>
       </Group>
     </Wrapper>