+++ assets/javascripts/kitten/karl/pages/homepage-2020/stories.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import HomePage from './index'
+
+export default {
+  component: HomePage,
+  title: 'pages/Homepage 2020',
+  parameters: {
+    component: HomePage,
+  },
+}
+
+export const Homepage = () => <HomePage />