+++ assets/javascripts/kitten/karl/pages/homepage-2020/text-with-side-image/stories.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import { TextWithSideImage } from './index'
+import { Container } from 'kitten'
+
+export default {
+  component: TextWithSideImage,
+  title: 'pages/Homepage 2020/TextWithSideImage',
+  parameters: {
+    component: TextWithSideImage,
+  },
+}
+
+export const DefaultProps = () => (
+  <Container>
+    <TextWithSideImage />
+  </Container>
+)