+++ assets/javascripts/kitten/karl/pages/homepage-2020/video/stories.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import { Video } from './index'
+import { Container } from 'kitten'
+
+export default {
+  component: Video,
+  title: 'pages/Homepage 2020/Video',
+  parameters: {
+    component: Video,
+  },
+}
+
+export const DefaultProps = () => (
+  <Container>
+    <Video />
+  </Container>
+)